Bug 786234 - Part 3: Plumbing for filtration, and bundle usage detection. r=abr

+++ media/webrtc/signaling/src/peerconnection/PeerConnectionImpl.cpp
@@ -1993,9 +1993,11 @@ PeerConnectionImpl::BuildStatsQuery_m(
     if (temp.get()) {
       query->streams.AppendElement(temp);
     } else {
-       CSFLogError(logTag, "Failed to get NrIceMediaStream for level %u "
+       CSFLogError(logTag, "Failed to get NrIceMediaStream for level %zu "
                            "in %s:  %s",
-                           uint32_t(level), __FUNCTION__, mHandle.c_str());
+                           static_cast<size_t>(level),
+                           __FUNCTION__,
+                           mHandle.c_str());
        MOZ_CRASH();
     }
   }
